Remove unnecessary prediction functions and files; slight refactoring
